pyproject.toml
[project]
name = "simple-twitter-mcp-server"
version = "0.1.0"
description = "Simple MCP server for Twitter scraping with Apify"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "apify-client>=1.7.0",
    "python-dotenv>=1.0.0",
    "weave>=0.52.7",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
